% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkColnames.R
\name{checkColnames}
\alias{checkColnames}
\title{Check Column Names}
\usage{
checkColnames(userSpecifiedColnames = NULL)
}
\arguments{
\item{userSpecifiedColnames}{the column names template file containing user specified column names for the input data. This file}
}
\value{
a validated user specified column names data set
}
\description{
Auxiliary function to check colnames for the input data set.
}
\author{
André Vidas Olsen
}
