% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactOutput_pmolCalc.R
\name{compactOutput_pmolCalc}
\alias{compactOutput_pmolCalc}
\title{Make Compact Output Containing Pico Mol Calculations}
\usage{
compactOutput_pmolCalc(data)
}
\description{
Make Compact Output Containing Pico Mol Calculations
}
\author{
André Vidas Olsen
}
