% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDataSet.R
\name{filterDataSet}
\alias{filterDataSet}
\title{Filter Data set}
\usage{
filterDataSet(data, endogene_lipid_db, ISTD_lipid_db,
  userSpecifiedColnames = NULL)
}
\arguments{
\item{data}{data formatted by the use of the mergeDataSet function from LipidQuan.}

\item{endogene_lipid_db}{the endogene lipid database}

\item{ISTD_lipid_db}{the ISTD lipid database}

\item{userSpecifiedColnames}{the column names template file containing user specified column names for the input data.}
}
\description{
filterDataSet selects relevants columns from a data set and remove species
if the name does not exist in the global database.
}
\author{
André Vidas Olsen
}
