% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColnames.R
\name{getColnames}
\alias{getColnames}
\title{Get Column Names}
\usage{
getColnames(userSpecifiedColnames = NULL)
}
\arguments{
\item{userSpecifiedColnames}{the column names template file containing user specified column names for the input data. This file}
}
\description{
Auxiliary function to get colnames for the input data set.
}
\author{
André Vidas Olsen
}
