% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDatabase.R
\name{makeDatabase}
\alias{makeDatabase}
\title{Make Database}
\usage{
makeDatabase(userSpecifiedColnames = NULL, DB_type)
}
\arguments{
\item{userSpecifiedColnames}{the column names template file containing user specified column names for the input data. This file}

\item{DB_type}{a string describing which of the two databases that is should be reset. "endo" for the endogene_lipid_db.csv database and "ISTD" for the ISTD_lipid_db.csv database.}
}
\description{
This function creates a database based on the user specified column names template
}
\author{
André Vidas Olsen
}
