% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeIndex_OH_DB_C.R
\name{makeIndex_OH_DB_C}
\alias{makeIndex_OH_DB_C}
\title{Creation of OH Index, DB Index and C-chain Index}
\usage{
makeIndex_OH_DB_C(data, userSpecifiedColnames = NULL)
}
\arguments{
\item{data}{data formatted by the use of the mergeDataSet function from LipidQuan.}

\item{userSpecifiedColnames}{the column names template file containing user specified column names for the input data.}
}
\description{
This function creates OH Index, DB Index and C-chain Index of data containing calculated %pmol values of samples.
}
\author{
André Vidas Olsen
}
