% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTableauOutput.R
\name{makeTableauOutput}
\alias{makeTableauOutput}
\title{Make Tableau Output}
\usage{
makeTableauOutput(classPmol_molPctClass, pmolCalculatedDataSet)
}
\arguments{
\item{classPmol_molPctClass}{BLA BLA}

\item{pmolCalculatedDataSet}{BLA BLA}
}
\description{
This function creates an output file of the results in a format that can be used by Tableau.
}
\author{
André Vidas Olsen
}
