% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_endo_and_ISTD_db.R
\name{merge_endo_and_ISTD_db}
\alias{merge_endo_and_ISTD_db}
\title{Merge Endogene and ISTD Database}
\usage{
merge_endo_and_ISTD_db(endogene_lipid_db, ISTD_lipid_db)
}
\arguments{
\item{endogene_lipid_db}{the endogene lipid database}

\item{ISTD_lipid_db}{the ISTD lipid database}
}
\value{
a data set consisting of endogene and ISTD database merged together
}
\description{
This function merges the two components of the reference databaes together: the endogene lipid database and the ISTD lipid database.
}
\author{
André Vidas Olsen
}
