% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDataSets.R
\name{mergeDataSets}
\alias{mergeDataSets}
\title{Merge Data Sets}
\usage{
mergeDataSets(dataList, database, userSpecifiedColnames = NULL,
  multiply = NULL, list = NULL)
}
\description{
This function merges multiple data sets together. The data sets are listed in the dataList parameter.
}
\author{
André Vidas Olsen
}
