% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFinalOutputs.R
\name{mergeFinalOutputs}
\alias{mergeFinalOutputs}
\title{Merge Final Outputs}
\usage{
mergeFinalOutputs(dataList)
}
\arguments{
\item{dataList}{a list of paths referring to input data}
}
\value{
a data set consisting of merged final output files
}
\description{
This function merges final output files into one data file
}
\author{
André Vidas Olsen
}
