% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap}
\usage{
plotHeatmap(data, controlStart, caseStart, filter = FALSE, threshold_logFC,
  threshold_amountOfInsignificantLogFC)
}
\arguments{
\item{data}{bla bla}

\item{controlStart}{bla bla}

\item{caseStart}{bla bla}
}
\description{
This function plots heatmaps of the data
}
\author{
André Vidas Olsen
}
