% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot Heatmap}
\usage{
plotPCA(data, groups, pathToOutput)
}
\arguments{
\item{data}{quantified data to be visualized (finalOutput_molPct.csv)}

\item{groups}{file that associates column names of MS1 with groups}

\item{pathToOutput}{the directory path to save the plots}
}
\description{
This function plots the following PCA plots: screeplot and biplot
}
\author{
André Vidas Olsen
}
