% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC_totalLipids.R
\name{plotQC_totalLipids}
\alias{plotQC_totalLipids}
\title{picomol Calculation}
\usage{
plotQC_totalLipids(data, userSpecifiedColnames = NULL, pathToOutput)
}
\arguments{
\item{data}{data formatted by the use of the compactOutput_pmolCalc function from LipidQuan.}

\item{userSpecifiedColnames}{the column names template file containing user specified column names for the input data.}

\item{pathToOutput}{the directory path to save the plots}
}
\description{
This function creates QC plots of MS1 intensity data
}
\author{
André Vidas Olsen
}
