% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFile.R
\name{readFile}
\alias{readFile}
\title{Read File}
\usage{
readFile(dataPath)
}
\arguments{
\item{dataPath}{a path referring to input data}
}
\description{
Auxiliary function that reads a data set by using the destination stored in the dataPath parameter.
}
\author{
André Vidas Olsen
}
