% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmSpaceInBeginning.R
\name{rmSpaceInBeginning}
\alias{rmSpaceInBeginning}
\title{Remove Space In Beginning of NAME and SPECIE Rows}
\usage{
rmSpaceInBeginning(data, userSpecifiedColnames = NULL)
}
\arguments{
\item{data}{data formatted by the use of the mergeDataSet function from LipidQuan.}

\item{userSpecifiedColnames}{the column names template file containing user specified column names for the input data. This file}
}
\value{
data set
}
\description{
Auxiliary function that removes the first letter in NAME and SPECIE, if the word begins with a [SPACE]. It is used since, by experience, some rows begins with a space for these two columns which can results in errors.
}
\author{
André Vidas Olsen
}
