% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLipidQ.R
\name{runLipidQ}
\alias{runLipidQ}
\title{Lipidomics Analysis Tool Graphical User Interface}
\usage{
runLipidQ()
}
\value{
start up lipidQ graphical user interface
}
\description{
runLipidQ starts the graphical user interface (GUI) of
Lipidomics Analysis Tool. From this GUI, the user can specify different types
of procedures, input data, different parameter as well as the output
directory.
}
\examples{
# run lipidQ
#runLipidQ()
}
\author{
André Vidas Olsen
}
