% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLipidQuan.R
\name{runLipidQuan}
\alias{runLipidQuan}
\title{Lipidomics Analysis Tool Graphical User Interface}
\usage{
runLipidQuan()
}
\arguments{
\item{testMode}{logical argument to define whether or not the function is
used for testing purposes.}
}
\value{
start up lipidQuan graphical user interface
}
\description{
runLipidQuan starts the graphical user interface (GUI) of
Lipidomics Analysis Tool. From this GUI, the user can specify different types
of procedures, input data, different parameter as well as the output
directory.
}
\examples{
# run lipidQuan, remove the testMode arguemnt when using the runLipidQuan
# function.

# runLipidQuan()
}
\author{
André Vidas Olsen
}
