% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setColnames.R
\name{setColnames}
\alias{setColnames}
\title{Set Column Names}
\usage{
setColnames(userSpecifiedColnames = NULL)
}
\description{
Auxiliary function to set colnames for the input data set. If no user specified column names is provided as an argument, default column names will be used
}
\author{
André Vidas Olsen
}
