% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactOutput_pmolCalc.R
\name{compactOutput_pmolCalc}
\alias{compactOutput_pmolCalc}
\title{Make Compact Output Containing Pico Mol Calculations}
\usage{
compactOutput_pmolCalc(data, userSpecifiedColnames = NULL)
}
\arguments{
\item{data}{data formatted by the use of the mergeDataSet function from LipidQuan.}

\item{userSpecifiedColnames}{the column names template file containing user specified column names for the input data.}
}
\description{
compactOutput_pmolCalc saves a data.frame with only
NAME, CLASS_PMOL_SUBT_PMOL_*, and MOL_PCT_CLASS_SUBT_PMOL_* and FILTERED
}
\author{
André Vidas Olsen
}
