% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactOutput_pmolCalc.R
\name{compactOutput_pmolCalc}
\alias{compactOutput_pmolCalc}
\title{Make Compact Output Containing Pico Mol Calculations}
\usage{
compactOutput_pmolCalc(data)
}
\description{
compactOutput_pmolCalc saves a data.frame with only
NAME, CLASS_PMOL_SUBT_PMOL_PREC*, and MOL_PCT_CLASS_SUBT_PMOL_PREC* OMSKRIV!!!
}
\author{
André Vidas Olsen
}
