% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactOutput_pmolCalc.R
\name{compactOutput_pmolCalc}
\alias{compactOutput_pmolCalc}
\title{Make Compact Output Containing Pico Mol Calculations}
\usage{
compactOutput_pmolCalc(data, userSpecifiedColnames = NULL)
}
\arguments{
\item{data}{data formatted by the use of the mergeDataSet function from LipidQuan.}

\item{userSpecifiedColnames}{the column names template file containing user specified column names for the input data.}
}
\value{
a compact data output comprising of pmol and pmol related columns
}
\description{
compactOutput_pmolCalc saves a compact version of data.frame
NAME, CLASS_PMOL_SUBT_PMOL_*, and MOL_PCT_CLASS_SUBT_PMOL_* and FILTERED
}
\examples{
# load user specified column names file.
userSpecifiedColnames <- read.table(system.file("extdata/LipidQ_DataBase", "userSpecifiedColnames.csv", package = "lipidQuan"), stringsAsFactors = FALSE, header = TRUE, sep = ",")

# load pmolCalculatedDataSet.csv made by using the pmolCalc() function
pmolCalculatedDataSet <- read.table(system.file("extdata/dataTables", "pmolCalculatedDataSet.csv", package = "lipidQuan"), stringsAsFactors = FALSE, header = TRUE, sep = ",")

# make compact output from pmolCalculatedDataSet
compact_output <- compactOutput_pmolCalc(pmolCalculatedDataSet, userSpecifiedColnames = userSpecifiedColnames)
}
\author{
André Vidas Olsen
}
