% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDataSet.R
\name{filterDataSet}
\alias{filterDataSet}
\title{Filter Data set}
\usage{
filterDataSet(data, database, userSpecifiedColnames = NULL)
}
\description{
filterDataSet selects relevants columns from a data set and remove species
if the name does not exist in the global database.
}
\author{
André Vidas Olsen
}
