% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColnames.R
\name{getColnames}
\alias{getColnames}
\title{Get Column Names}
\usage{
getColnames(userSpecifiedColnames = NULL)
}
\description{
Auxiliary function to get colnames for the input data set. If no user specified column names is provided as an argument, default column names will be returned.
}
\author{
André Vidas Olsen
}
