% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDatabase.R
\name{makeDatabase}
\alias{makeDatabase}
\title{Make Database}
\usage{
makeDatabase(userSpecifiedColnames = NULL, DB_type)
}
\arguments{
\item{userSpecifiedColnames}{the column names template file containing user
specified column names for the input data. This file}

\item{DB_type}{a string describing which of the two databases that is should
be reset. "endo" for the endogene_lipid_db.csv database and "ISTD" for the
ISTD_lipid_db.csv database.}
}
\value{
a database template file in the form of either endogene or ISTD
}
\description{
This function creates a database based on the user specified
column names template
}
\examples{
# load user specified column names files
 userSpecifiedColnames <- read.table(system.file("extdata/LipidQ_DataBase",
 "userSpecifiedColnames.csv", package = "lipidQ"),
 stringsAsFactors = FALSE, header = TRUE, sep = ",")

# make endo database
makeDatabase(userSpecifiedColnames = userSpecifiedColnames, DB_type = "endo")

# make ISTD database
makeDatabase(userSpecifiedColnames = userSpecifiedColnames, DB_type = "ISTD")
}
\author{
André Vidas Olsen
}
