% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeIndex_OH_DB_C.R
\name{makeIndex_OH_DB_C}
\alias{makeIndex_OH_DB_C}
\title{Creation of OH Index, DB Index and C-chain Index}
\usage{
makeIndex_OH_DB_C(data, userSpecifiedColnames = NULL)
}
\arguments{
\item{data}{data formatted by the use of the mergeDataSet function from LipidQuan.}

\item{userSpecifiedColnames}{the column names template file containing user specified column names for the input data.}
}
\value{
a data set consisting of OH, DB, C indexes
}
\description{
This function creates OH Index, DB Index and C-chain Index of data containing calculated %pmol values of samples.
}
\examples{
# load user specified column names file.
userSpecifiedColnames <- read.table(system.file("extdata/LipidQ_DataBase", "userSpecifiedColnames.csv", package = "lipidQuan"), stringsAsFactors = FALSE, header = TRUE, sep = ",")

# load pmolCalculatedDataSet.csv made by using the pmolCalc() function
pmolCalculatedDataSet <- read.table(system.file("extdata/dataTables", "pmolCalculatedDataSet.csv", package = "lipidQuan"), stringsAsFactors = FALSE, header = TRUE, sep = ",")

# make data of OH, DB and C indexes
indexData <- makeIndex_OH_DB_C(pmolCalculatedDataSet, userSpecifiedColnames = userSpecifiedColnames)

# print data
indexData[[1]] # OH index
indexData[[2]] # DB index
indexData[[3]] # C index
}
\author{
André Vidas Olsen
}
