% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDataSets.R
\name{mergeDataSets}
\alias{mergeDataSets}
\title{Merge Data Sets}
\usage{
mergeDataSets(dataList, database, userSpecifiedColnames = NULL,
  correctionList = NULL, multiply = NULL)
}
\arguments{
\item{dataList}{a list of paths referring to input data}

\item{database}{a database containing reference data for the input data, e.g. classes of interest, QUAN_MODE, QUAN_SCAN etc.}

\item{userSpecifiedColnames}{the column names template file containing user specified column names for the input data. This file
is used to translate the user specified column names to the program, so that it uses the correct columns for the different analysis procedures.}

\item{correctionList}{a file containing a list of sum compositions to be multiplied for in the MS1 column values (intensity values)}

\item{multiply}{a parameter used to multiply intensity values in the MS1 column values if the corresponding sum composition appear in a correction list file used as argument in the list parameter FORKLAR MENINGEN}
}
\description{
This function merges multiple data sets together. The data sets are listed in the dataList parameter.
}
\author{
André Vidas Olsen
}
