% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDataSets.R
\name{mergeDataSets}
\alias{mergeDataSets}
\title{Merge Data Sets}
\usage{
mergeDataSets(dataList, endogene_lipid_db, ISTD_lipid_db,
  userSpecifiedColnames = NULL, correctionList = NULL, multiply = NULL)
}
\arguments{
\item{dataList}{a list of paths referring to input data}

\item{endogene_lipid_db}{the endogene lipid database}

\item{ISTD_lipid_db}{the ISTD lipid database}

\item{userSpecifiedColnames}{the column names template file containing user
specified column names for the input data.
is used to translate the user specified column names to the program, so that
it uses the correct columns for the different analysis procedures.}

\item{correctionList}{a file containing a list of sum compositions to be
multiplied for in the MS1 column values (intensity values)}

\item{multiply}{a parameter used to multiply intensity values in the MS1
column on selected sum compositions. The parameter is useful if lipidX is
used to obtain the intensity data derived from overlapping MS scan ranges.
The sum composition are selected by the user and should appear in a
correction list file that is used as argument for the correctionList
parameter.}
}
\value{
a data set consisting of merged input data
}
\description{
This function merges multiple data sets together. The data sets
are listed in the dataList parameter.
}
\author{
André Vidas Olsen
}
