% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFinalOutputs.R
\name{mergeFinalOutputs}
\alias{mergeFinalOutputs}
\title{Merge Final Outputs}
\usage{
mergeFinalOutputs(dataList)
}
\arguments{
\item{dataList}{a list of paths referring to input data}
}
\value{
a data set consisting of merged final output files
}
\description{
This function merges final output molPct files into one data
file
}
\examples{
# make output molPct files data path list
dataPathList <- c(system.file("extdata/finalOutputFiles",
                  "finalOutput1.csv", package = "lipidQuan"),
                 system.file("extdata/finalOutputFiles",
                  "finalOutput2.csv", package = "lipidQuan"))

# merge final output files to one file
mergeFinalOutputs(dataPathList)
}
\author{
André Vidas Olsen
}
