% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTableauDataSets.R
\name{mergeTableauDataSets}
\alias{mergeTableauDataSets}
\title{Merge Tableau Data Sets}
\usage{
mergeTableauDataSets(dataList)
}
\arguments{
\item{dataList}{a list of paths referring to input data}
}
\description{
This function merges tableau output files into one data file
}
\author{
André Vidas Olsen
}
