% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap}
\usage{
plotHeatmap(data, sampleTypes, K = NULL, pathToOutput, log2 = FALSE,
  pseudoCount = NULL)
}
\arguments{
\item{data}{quantified data to be visualized (finalOutput_molPct.csv)}

\item{sampleTypes}{file that associates column names of MS1 with the sample types}

\item{K}{number of K in the Kmeans algorithm.}

\item{pathToOutput}{the directory path to save the plots}

\item{log2}{logical argument that specifies whether or not data has to be log2 transformed}

\item{pseudoCount}{pseudo count added to the data if the data is log2 transformed in order to avoid negative infinite values in the data}
}
\value{
Two heatmap images: heatmap of species (heatmapSpecies_K_n.png) and heatmap of classes (heatmapClasses_K_n.png), where n indicates the amount of clusters K. If clusters are not chosen by the user, a majority vote chart will also be included called nbclustResults.png.
}
\description{
This function plots heatmaps of the data
}
\examples{
# load sample types file and data to be used for visualization
sampleTypes <- read.csv(system.file("extdata", "sampleTypes.csv", package = "lipidQuan"), stringsAsFactors = FALSE)
data <- read.csv(system.file("extdata/dataTables", "finalOutput_molPct.csv", package = "lipidQuan"), stringsAsFactors = FALSE)


# create heatmap from log2 transformed data with 0.0001 added pseudo counts
plotHeatmap(data = data, sampleTypes = sampleTypes, pathToOutput = system.file("extdata/dataTables", package = "lipidQuan"), log2 = TRUE, pseudoCount = 0.0001)
}
\author{
André Vidas Olsen
}
