% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap}
\usage{
plotHeatmap(data, groups, K = NULL, pathToOutput)
}
\arguments{
\item{data}{quantified data to be visualized (finalOutput_molPct.csv)}

\item{groups}{file that associates column names of MS1 with groups}

\item{K}{number of K in the Kmeans algorithm.}

\item{pathToOutput}{the directory path to save the plots}
}
\description{
This function plots heatmaps of the data
}
\author{
André Vidas Olsen
}
