% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap}
\usage{
plotHeatmap(data, sampleTypes, K = NULL, pathToOutput, log2 = FALSE,
  pseudoCount = NULL)
}
\arguments{
\item{data}{quantified data to be visualized (finalOutput_molPct.csv)}

\item{sampleTypes}{file that associates column names of MS1 with the sample types}

\item{K}{number of K in the Kmeans algorithm.}

\item{pathToOutput}{the directory path to save the plots}

\item{log2}{logical argument that specifies whether or not data has to be log2 transformed}

\item{pseudoCount}{pseudo count added to the data if the data is log2 transformed in order to avoid negative infinite values in the data}
}
\description{
This function plots heatmaps of the data
}
\author{
André Vidas Olsen
}
