% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot Heatmap}
\usage{
plotPCA(data, sampleTypes, pathToOutput, log2 = FALSE, pseudoCount = NULL)
}
\arguments{
\item{data}{quantified data to be visualized (finalOutput_molPct.csv)}

\item{sampleTypes}{file that associates column names of MS1 with sampleTypes}

\item{pathToOutput}{the directory path to save the plots}

\item{log2}{logical argument that specifies whether or not data has to be log2 transformed}

\item{pseudoCount}{pseudo count added to the data if the data is log2 transformed in order to avoid negative infinite values in the data}
}
\description{
This function plots the following PCA plots: screeplot and biplot
}
\author{
André Vidas Olsen
}
