% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC_ISTD.R
\name{plotQC_ISTD}
\alias{plotQC_ISTD}
\title{picomol Calculation}
\usage{
plotQC_ISTD(data, endogene_lipid_db, ISTD_lipid_db,
  userSpecifiedColnames = NULL, pathToOutput, blnkReplicates = FALSE,
  numberOfReplicates)
}
\arguments{
\item{data}{data formatted by the use of the mergeDataSet function from LipidQuan.}

\item{endogene_lipid_db}{the endogene lipid database}

\item{ISTD_lipid_db}{the ISTD lipid database}

\item{userSpecifiedColnames}{the column names template file containing user specified column names for the input data.}

\item{pathToOutput}{the directory path to save the plots}

\item{blnkReplicates}{logical parameter for specifying whether the blank sample contains replicates or not. FALSE: no replicates, TRUE: replicates.}

\item{numberOfReplicates}{the number of replicates for each sample}
}
\description{
This function creates QC plots of MS1 intensity data
}
\author{
André Vidas Olsen
}
