% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmolCalc.R
\name{pmolCalc}
\alias{pmolCalc}
\title{picomol Calculation}
\usage{
pmolCalc(data, endogene_lipid_db, ISTD_lipid_db, userSpecifiedColnames = NULL,
  spikeISTD, zeroThresh)
}
\arguments{
\item{data}{data formatted by the use of the mergeDataSet function from LipidQuan.}

\item{endogene_lipid_db}{the endogene lipid database}

\item{ISTD_lipid_db}{the ISTD lipid database}

\item{userSpecifiedColnames}{the column names template file containing user specified column names for the input data.}

\item{spikeISTD}{internal standard spike amount in uL}

\item{zeroThresh}{an optional threshold that determines if a given small value in mol pct. specie composition columns should be rounded down to zero.}
}
\description{
This function calculates pico mol (pmol) of species based on intensity from measurements (target specie + internal standard) and known quantity of internal standard
}
\author{
André Vidas Olsen
}
