% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmolCalc.R
\name{pmolCalc}
\alias{pmolCalc}
\title{picomol Calculation}
\usage{
pmolCalc(data, database, userSpecifiedColnames = NULL, spikeVariable,
  zeroThresh)
}
\description{
This function calculates pico mol (pmol) of species based on intensity from measurements (target specie + internal standard) and known quantity of internal standard
}
\author{
André Vidas Olsen
}
