% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_is.R
\name{sort_is}
\alias{sort_is}
\title{Sort Internal Standards}
\usage{
sort_is(data, userSpecifiedColnames = NULL)
}
\arguments{
\item{data}{data formatted by the use of the mergeDataSet function from LipidQuan.}

\item{userSpecifiedColnames}{the column names template file containing user specified column names for the input data. This file}
}
\description{
This function orders all internal standards (is) in a given data set so that they appear as the last rows
(after the normal species). The purpose for this is mainly to get a better overview of the data types when manually inspecting the data.
}
\author{
André Vidas Olsen
}
